azure_functions/requirements.txt
azure-functions
azure-identity
msgraph-sdk
office365-REST-python-client
openai
python-dotenv
